Add saifi report

+++ OMS.Web/OMS.Web.UI.Models/BindingModels/ReportOptions.cs
@@ -5,7 +5,7 @@
     public class ReportOptions
     {
         public ReportType Type { get; set; }
-        public int? ElementId { get; set; }
+        public long? ElementId { get; set; }
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
     }
