+++ OMS.Web/OMS.Web.UI.Models/ViewModels/ReportViewModel.cs
@@ -4,7 +4,7 @@
 
     public class ReportViewModel : IViewModel
     {
-        public IDictionary<string, int> Data { get; set; }
+        public IDictionary<string, float> Data { get; set; }
         public string Type { get; set; }
     }
 }
