+++ OutageManagementSystem/Common/OutageService/OutageReport.cs
@@ -4,7 +4,7 @@ namespace Outage.Common.OutageService
 {
     public class OutageReport
     {
-        public IDictionary<string, int> Data { get; set; }
+        public IDictionary<string, float> Data { get; set; }
         public string Type { get; set; }
     }
 }
